+++ src/Aevatar.Domain/Workflow/WorkflowUintInfo.cs
@@ -4,6 +4,6 @@ public class WorkflowUintInfo
 {
     public string GrainId { get; set; }
     public string? NextGrainId { get; set; }
-    public int XPosition { get; set; }
-    public int YPosition { get; set; }
+    public float XPosition { get; set; }
+    public float YPosition { get; set; }
 }
